Move mocks to a CommProtocolMocks target
Keeps throwaway-key mock factories out of the production CommProtocol library;
consumers opt in via import CommProtocolMocks.
