+++ Sources/CommProtocol/IdentityExchange/AppWelcome.swift
@@ -127,34 +127,3 @@ extension AppWelcome {
 		)
 	}
 }
-
-extension AppWelcome {
-	static public func mock(
-		remoteAgentKey: AgentPublicKey,
-		keyPackageData: Data
-	) throws -> AppWelcome {
-		let (identityKey, signedIdentity) =
-			try Mocks
-			.mockIdentity()
-
-		let groupId = DataIdentifier(width: .bits256)
-
-		let (agentKey, introduction) =
-			try identityKey
-			.createNewDelegate(
-				signedIdentity: signedIdentity,
-				identityMutable: .mock(),
-				agentType: .welcome(
-					remoteAgentId: remoteAgentKey,
-					groupId: groupId
-				)
-			)
-
-		return try agentKey.createAppWelcome(
-			introduction: introduction,
-			agentData: .mock(),
-			groupId: groupId,
-			keyPackageData: keyPackageData
-		)
-	}
-}
