+++ Tests/CommProtocolTests/Anchors/AnchorAPIs.swift
@@ -8,6 +8,7 @@
 import AtprotoTypes
 import AtprotoTypesMocks
 import CommProtocol
+import CommProtocolMocks
 import CryptoKit
 import Testing
