+++ Tests/CommProtocolTests/CommProtocolTests.swift
@@ -6,6 +6,7 @@
 //
 
 import CommProtocol
+import CommProtocolMocks
 import CryptoKit
 import Foundation
 import Testing
